pyproject.toml
[project]
name = "f1-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jwt>=1.4.0",
    "mcp[cli]>=1.14.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
